extracted wmts functions
